+++ NEWS.md
@@ -1,6 +1,7 @@
 # random.cdisc.data 0.3.12.9007
 * Fixed a typo in `ADSL$BMRKR1`'s label.
-* Added new random dataset generator: PK Analysis Dataset (`radpc`) .
+* Added new random dataset generator: PK Analysis Dataset (`radpc`).
+* Added new random dataset generator: PK Parameter Dataset (`radpp`).
 
 ### Miscellaneous
 * Updated R version requirement to >= 3.6.